// src/services/rest-api/customer-service/model/personal-info-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.242
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PersonalInfoModel
 */
export interface PersonalInfoModel {
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'titleCodeTh'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'titleNameTh'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'firstNameTh'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'lastNameTh'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'titleCodeEn'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'titleNameEn'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'firstNameEn'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'lastNameEn'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'genderCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'birthDate'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'nationalityCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'identityType'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'identityNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'identityExpireDate'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'occupationCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'jobTypeCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'jobPosition'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof PersonalInfoModel
     */
    'income'?: number;
    /**
     * 
     * @type {number}
     * @memberof PersonalInfoModel
     */
    'incomeOther'?: number;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'incomeSourceOther'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoModel
     */
    'incomeSourceOversea'?: string | null;
}