// src/services/rest-api/customer-service/model/personal-info-response-data-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.242
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { PersonalInfoResponse } from './personal-info-response';

/**
 * 
 * @export
 * @interface PersonalInfoResponseDataResponse
 */
export interface PersonalInfoResponseDataResponse {
    /**
     * 
     * @type {number}
     * @memberof PersonalInfoResponseDataResponse
     */
    'status'?: number;
    /**
     * 
     * @type {{ [key: string]: Array<string>; }}
     * @memberof PersonalInfoResponseDataResponse
     */
    'errors'?: { [key: string]: Array<string>; } | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoResponseDataResponse
     */
    'message'?: string | null;
    /**
     * 
     * @type {PersonalInfoResponse}
     * @memberof PersonalInfoResponseDataResponse
     */
    'data'?: PersonalInfoResponse | null;
}